refactor selection into own file
