+++ app/packages/core/src/components/Schema/SchemaSettings.tsx
@@ -3,8 +3,8 @@ import styled from "styled-components";
 
 import { Box } from "@mui/material";
 import CloseIcon from "@mui/icons-material/Close";
-import Checkbox from "@mui/material/Checkbox";
-import { Button, useTheme } from "@fiftyone/components";
+
+import { useTheme } from "@fiftyone/components";
 import { TabOption } from "../utils";
 
 import useSchemaSettings, {
@@ -13,6 +13,7 @@ import useSchemaSettings, {
 } from "@fiftyone/state/src/hooks/useSchemaSettings";
 
 import { SchemaSearch } from "./SchemaSearch";
+import { SchemaSelection } from "./SchemaSelection";
 
 const ModalWrapper = styled.div`
   position: fixed;
@@ -51,7 +52,6 @@ interface Props {
 }
 
 const SchemaSettings = (props: Props) => {
-  const { test } = props;
   const theme = useTheme();
 
   const schemaModalRef = useRef<HTMLDivElement>(null);
@@ -61,21 +61,14 @@ const SchemaSettings = (props: Props) => {
     setSettingsModal,
     searchTerm,
     setSearchTerm,
-    finalSelectedPaths,
     setSelectedTab,
     originalSelectedPaths,
     setSelectedPaths,
     selectedTab,
-    fieldsOnly,
-    setFieldsOnly,
-    setView,
-    toggleSelection,
-    finalSchema,
-    allFieldsChecked,
-    setAllFieldsChecked,
   } = useSchemaSettings();
 
   const { open: isSettingsModalOpen } = settingModal || {};
+  console.log("isSettingsModalOpen", isSettingsModalOpen);
 
   if (!isSettingsModalOpen) {
     return null;
@@ -152,178 +145,7 @@ const SchemaSettings = (props: Props) => {
               searchTerm={searchTerm}
             />
           )}
-          {selectedTab === TAG_OPTIONS_MAP.SELECTION && (
-            <Box
-              display="flex"
-              flexDirection="column"
-              sx={{ position: "relative !important" }}
-            >
-              <Box
-                style={{
-                  position: "relative",
-                  padding: "0.5rem 0",
-                  color: theme.text.primary,
-                  display: "flex",
-                }}
-              >
-                Show attributes
-                <Checkbox
-                  name={"Carousel"}
-                  value={!fieldsOnly}
-                  checked={!fieldsOnly}
-                  onChange={() => setFieldsOnly(!fieldsOnly)}
-                  style={{ padding: "4px" }}
-                />
-              </Box>
-              {!allFieldsChecked && (
-                <Box
-                  style={{
-                    position: "relative",
-                    color: theme.text.primary,
-                    display: "flex",
-                  }}
-                >
-                  Select All
-                  <Checkbox
-                    name={"Carousel"}
-                    value={allFieldsChecked}
-                    checked={allFieldsChecked}
-                    onChange={() => setAllFieldsChecked(!allFieldsChecked)}
-                    style={{ padding: "4px" }}
-                  />
-                </Box>
-              )}
-              {allFieldsChecked && (
-                <Box
-                  style={{
-                    position: "relative",
-                    color: theme.text.primary,
-                    display: "flex",
-                  }}
-                >
-                  Deselect All
-                  <Checkbox
-                    name={"Carousel"}
-                    value={allFieldsChecked}
-                    checked={allFieldsChecked}
-                    onChange={() => setAllFieldsChecked(!allFieldsChecked)}
-                    style={{ padding: "4px" }}
-                  />
-                </Box>
-              )}
-              <Box
-                style={{
-                  position: "relative",
-                  height: "50vh",
-                  marginTop: "1rem",
-                  overflow: "auto",
-                  color: "#232323",
-                  border: `1px solid ${theme.primary.plainBorder}`,
-                }}
-              >
-                {finalSchema.map((item) => {
-                  const {
-                    path,
-                    count,
-                    isSelected,
-                    pathLabelFinal,
-                    skip,
-                    disabled,
-                  } = item;
-
-                  if (skip) return null;
-
-                  return (
-                    <Box
-                      style={{
-                        padding: "0.25rem 0.25rem",
-                        borderBottom: `1px solid ${theme.primary.plainBorder}`,
-                        display: "flex",
-                      }}
-                      key={path}
-                    >
-                      <Box>
-                        <Checkbox
-                          name={"Carousel"}
-                          value={path}
-                          checked={isSelected}
-                          onChange={() => {
-                            toggleSelection(path, isSelected);
-                          }}
-                          style={{
-                            padding: 0,
-                          }}
-                          disabled={disabled}
-                        />
-                      </Box>
-                      <Box
-                        style={{
-                          paddingLeft: `${(count - 1) * 15 + 5}px`,
-                          color: disabled
-                            ? theme.text.tertiary
-                            : theme.text.primary,
-                        }}
-                      >
-                        {pathLabelFinal}
-                      </Box>
-                    </Box>
-                  );
-                })}
-              </Box>
-              <Box
-                style={{
-                  position: "relative",
-                  display: "flex",
-                  padding: "1rem 0.25rem",
-                }}
-              >
-                <Button
-                  style={{
-                    color: theme.text.primary,
-                    marginRight: "0.5rem",
-                    boxShadow: "none",
-                    padding: "0.5rem 0.5rem",
-                    borderRadius: "4px",
-                  }}
-                  onClick={() => {
-                    const stageKwargs = [
-                      ["field_names", [...finalSelectedPaths]],
-                      ["_allow_missing", true],
-                    ];
-                    const stageCls = "fiftyone.core.stages.SelectFields";
-                    const stage = {
-                      _cls: stageCls,
-                      kwargs: stageKwargs,
-                    } as Stage;
-                    try {
-                      setView([stage]);
-                    } catch (e) {
-                      console.log("error", e);
-                    } finally {
-                      setSettingsModal({ open: false });
-                    }
-                  }}
-                >
-                  Add to view
-                </Button>
-                <Button
-                  style={{
-                    color: theme.text.primary,
-                    boxShadow: "none",
-                    padding: "0.5rem 0.5rem",
-                    borderRadius: "4px",
-                  }}
-                  onClick={() => {
-                    setSettingsModal({ open: false });
-                    setSearchTerm("");
-                    setSelectedPaths(originalSelectedPaths);
-                  }}
-                >
-                  cancel
-                </Button>
-              </Box>
-            </Box>
-          )}
+          {selectedTab === TAG_OPTIONS_MAP.SELECTION && <SchemaSelection />}
         </Container>
       </ModalWrapper>
     </Fragment>
